requirements.txt
numpy
pandas
streamlit
joblib
openpyxl
scikit-learn
plotly
Pillow